Update README.md
Moved leaderboard specific edits to top of file, updated some of the functions. Still need to explain some of the functions. New plan for next time is to get rid of the in depth function definitions (those are already in the code) and just aim for broad overstrokes/things necessary for setting up the system.