Update README.md
Completely got rid of the technical components of the Readme (I might throw them into a separate file later), and repurposed it as a high level guidance aimed at teachers who don't care about the specifics of the implementation. This involved giving an overall summary of what our leaderboard does and useful information for using it's settings. It is still a bit rough (especially since I copied the student Q@A because that has useful information in it), and so will need mostly formatting and some wording changes to finish it up.